Django/week9/patikagram/requirements.txt
asgiref==3.4.1
Django==4.0
django-extensions==3.1.5
djangorestframework==3.13.0
djangorestframework-simplejwt==5.0.0
Pillow==8.4.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
